modify 게시글 수정
Former-commit-id: d6a35b93eb853f2a31b999b06def2efd2f348167

+++ BackEnd/src/main/java/com/web/commitment/controller/BoardController.java
@@ -129,10 +129,11 @@ public class BoardController {
 	public int delete(@RequestParam String sns_id) {
 
 		Optional<Board> sns = boardDao.findById(sns_id);
-		System.out.println(sns);
 
 		try {
-			boardDao.delete(sns.get());
+			if(sns.isPresent()) 
+				boardDao.delete(sns.get());
+			
 			return 0;
 
 		} catch (Exception e) {
